email and user table and auto creation. New credentials added and env

+++ BlockChain_code_File-main/models/user.js
@@ -25,7 +25,7 @@ const User = db.define('user_table', {
     allowNull: false,
   },
   user_type:{
-    type: DataTypes.INTEGER,
+    type: DataTypes.INTEGER(255),
     allowNull:false,
     defaultValue:0
   }
